requirements.txt
prettytable==0.7.2
pytest==5.4.3
PyYAML==5.3.1

